Refactor exported types
